delete/exit confirmation models moved to components
