+++ client/src/components/DeleteConfirmationModal.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+/**
+ * DeleteConfirmationModal Component for confirming entry deletion
+ * 
+ * @param {Object} props - Component props
+ * @param {Function} props.onConfirm - Function to handle deletion confirmation
+ * @param {Function} props.onCancel - Function to handle cancellation
+ * @returns {JSX.Element} The confirmation modal component
+ */
+const DeleteConfirmationModal = ({ onConfirm, onCancel }) => {
+    return (
+        <div className="fixed inset-0 bg-black bg-opacity-80 flex items-center justify-center z-50">
+            <div className="bg-gray-900 bg-opacity-90 rounded-lg p-6 text-white border border-purple-500 shadow-xl">
+                <h2 className="text-xl font-bold mb-4 text-blue-400">Confirm Deletion</h2>
+                <p className="mb-6 text-gray-300">Are you sure you want to delete this entry? This action cannot be undone.</p>
+                <div className="flex justify-end space-x-4">
+                    <button
+                        onClick={onCancel}
+                        className="px-4 py-2 text-white bg-gray-600 hover:bg-gray-700 rounded-lg"
+                    >
+                        Cancel
+                    </button>
+                    <button
+                        onClick={onConfirm}
+                        className="px-4 py-2 text-white bg-red-600 hover:bg-red-700 rounded-lg border border-red-400"
+                    >
+                        Delete
+                    </button>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default DeleteConfirmationModal;
